+++ src/app/ui/report-view/analysis-summary.tsx
