+++ src/app/ui/report-view/diagram-box.tsx
