+++ src/app/ui/report-view/summary-box.tsx
